fix: lint

+++ src/functions/src/category/dao.js
@@ -1,11 +1,11 @@
-const db = require("../config/database");
+const db = require('../config/database');
 
 exports.selectCategoryDAO = async ({ id, transaction_type, name, user_id }) => {
   const response = await db.query(
     `SELECT * FROM categories WHERE users_fk = $1
-    ${id ? `and id = '${id}'` : ""}
-    ${transaction_type ? `and transaction_type = '${transaction_type}'` : ""}
-    ${name ? `and name = '${name}'` : ""}`,
+    ${id ? `and id = '${id}'` : ''}
+    ${transaction_type ? `and transaction_type = '${transaction_type}'` : ''}
+    ${name ? `and name = '${name}'` : ''}`,
     [user_id]
   );
   return response;
@@ -23,17 +23,11 @@ exports.insertCategoryDAO = async ({ transaction_type, name, user_id }) => {
 };
 
 exports.updateCategoryDAO = async ({ id, name }) => {
-  const response = await db.query(
-    "UPDATE categories SET name = $1 WHERE id = $2",
-    [name, id]
-  );
+  const response = await db.query('UPDATE categories SET name = $1 WHERE id = $2', [name, id]);
   return response;
 };
 
 exports.deleteCategoryDAO = async ({ id, user_id }) => {
-  const response = await db.query(
-    "DELETE FROM categories WHERE id = $1 and users_fk = $2",
-    [id, user_id]
-  );
+  const response = await db.query('DELETE FROM categories WHERE id = $1 and users_fk = $2', [id, user_id]);
   return response;
 };
